chore: Add SimulatorSettings model for ADB port configuration
