+++ src/types/simulator.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class SimulatorSettings(BaseModel):
+    adb_port: list[int]
